// nightrider/src/remediation-coordinator/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, Logger, MRNObject, Queue, ServiceStub, SmartBucket, SmartMemory, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CONVERSATION_MEMORY: SmartMemory;
  INCIDENT_DATA: ActorNamespace<import('../incident-data/index.js').IncidentData>;
  INCIDENT_ORCHESTRATOR: ServiceStub<import('../incident-orchestrator/index.js').default>;
  KNOWLEDGE_BASE: SmartBucket;
  logger: Logger;
  NOTIFICATION_QUEUE: Queue<import('../notification-observer').Body>;
  RCA_ENGINE: ServiceStub<import('../rca-engine/index.js').default>;
  SRE_API: ServiceStub<import('../sre-api/index.js').default>;
  tracer: Tracer;
}